Unqualify classes in too general type signature/annotation warnings
